Able to increase health points by pickups. Change health to float instead of int.

+++ Assets/Scripts/UIController.cs
@@ -24,7 +24,7 @@ public class UIController : MonoBehaviour
         scoreLabel.text = score.ToString();
     }
 
-    public void UpdateHealthDisplay(int health)
+    public void UpdateHealthDisplay(float health)
     {
         healthLabel.text = $"HP: {health}";
